fix: server host default is localhost
This is what we used to have until the default changed in af3353668e32a4b3c3fc746b94744327b5f2212c. However this is not an idiomatic host name in development for most users so reverting the change of host default. This also should fix the e2e tests which have broken since.

Still, we want environments where `0.0.0.0` is the required host to cause the default to become `0.0.0.0`. This change goes against that, and so we will re-open #324.

CC @jgillick