fix(schema): do not import array type in context typegen
closes #452

I attempted to resolve this issue using just the TS Compiler API but
after two or so hours had made no real progress with it. I even opened
up a Stack Overflow issue about it: https://stackoverflow.com/questions/60610543/how-can-i-use-the-typescript-compiler-api-to-extract-the-type-of-an-array. Alas I gave in and used ts-morph to get the job done in minutes...