improve: remove codename dep (#847)
closes #167

While this may not seem like an improvement for end-users, but rather an internal refactor, the user-facing part is in how it helps support #119. It turned out that codename uses some dynamic requires that rollup commonjs plugin cannot understand.

This does mean that users will see a less pretty name when creating new projects in non-empty directories. Very minor thing that we can revisit.